requirements.txt
# Google Earth Engine
earthengine-api>=1.4.0

# Raster processing
rasterio>=1.4.0
xarray>=2024.1.0
rioxarray>=0.17.0

# Scientific computing
numpy>=1.26.0
scipy>=1.12.0

# Visualization
matplotlib>=3.8.0

# Utilities
tqdm>=4.66.0
python-dotenv>=1.0.0
click>=8.1.0
requests>=2.31.0
